build(debug): remove docker and development/debug instructions
Could not get docker build as defined in Hubot Test Helper to work. May come back to this, but as
this project is only intended to run as a dev dependency of other apps, which themselves run on a
variety of node versions, I don't see the point in running/debugging it in a container.